// packages/hardhat/ts/types.ts
export interface IDeployedPoseidonContracts {
  PoseidonT3Contract: PoseidonT3;
  PoseidonT4Contract: PoseidonT4;
  PoseidonT5Contract: PoseidonT5;
  PoseidonT6Contract: PoseidonT6;
}

/**
 * An interface that represents the arguments for MACI contracts deployment.
 */
export interface IDeployMaciArgs {
  hre: HardhatRuntimeEnvironment;
  /**
   * The address of the SignUpTokenGatekeeper contract
   */
  signUpTokenGatekeeperContractAddress: string;

  /**
   * The address of the ConstantInitialVoiceCreditProxy contract
   */
  initialVoiceCreditBalanceAddress: string;

  /**
   * The address of the TopupCredit contract
   */
  topupCreditContractAddress: string;

  /**
   * The signer to use to deploy the contract
   */
  signer?: Signer;

  /**
   * Poseidon contract addresses (if not provided, they will be deployed automatically)
   */
  poseidonAddresses?: Partial<{
    poseidonT3: string;
    poseidonT4: string;
    poseidonT5: string;
    poseidonT6: string;
  }>;

  /**
   * The depth of the state tree
   */
  stateTreeDepth?: number;

  /**
   * Whether to suppress console output
   */
  quiet?: boolean;

  /**
   * Whether to support QV or not
   */
  useQv?: boolean;
}

/**
 * An interface that represents the deployed MACI contracts.
 */
export interface IDeployedMaci {
  maciContract: MACI;
  stateAqContract: AccQueueQuinaryMaci;
  pollFactoryContract: PollFactory;
  poseidonAddrs: {
    poseidonT3: string;
    poseidonT4: string;
    poseidonT5: string;
    poseidonT6: string;
  };
}